removed manual installation of 7z

+++ online/winget.ps1
@@ -20,7 +20,34 @@
 
 Add-AppxPackage -Path "D:\data\winget-cli.msixbundle"
 
-winget install 7zip.7zip --force --silent
+$checkPath = "C:\Program Files\7-Zip\7z.exe"
+$setupPath = Join-Path $scriptDirectory "apps\7z.msi"
+
+if (-not (Test-Path $checkPath)) {
+    if (-not (Test-Path $setupPath)) {
+        try {
+            winget install 7zip.7zip --force --silent
+            # $client = New-Object System.Net.WebClient
+            # "https://7-zip.org/a/.msi"
+            # $client.DownloadFile("https://7-zip.org/a/7z2301-x64.msi", $setupPath)
+            Write-Host "7-Zip setup file downloaded."
+        } catch {
+            Write-Host "An error occurred while downloading the 7-Zip setup file."
+        }
+    }
+
+    try {
+        Start-Process msiexec.exe -Wait -ArgumentList "/i `"$setupPath`" /quiet"
+        Write-Host "Installed 7-Zip"
+    } catch {
+        Write-Host "An error occurred while installing 7-Zip."
+    }
+} else {
+    Write-Host "7-Zip is already installed."
+}
+
+
 
 winget install Zoom.ZoomOutlookPlugin --force --silent
 
+# winget install Mozilla.Firefox --force --silent
